Use 8099 port in example and show alternative approach
Expose the default container port 8099 in the sample Docker Compose YAML file and add a separate note explaining how you can expose the container port on a different port if there is a conflicting service running on the same port. 

The original intention of exposing the container on port 80 was to make accessing the service more intuitive and to provide a 'self documenting' example of exposing the container port on a different port, but we received feedback that port 80 is commonly already taken so this could be confusing to people who are new to Docker and didn't know how to resolve the conflicting port.